+++ packages/f-registration/src/components/Registration.vue
@@ -3,7 +3,7 @@
         <bag-celebrate-icon :class="$style['bag-icon']" />
         <card
             :data-theme-registration="theme"
-            :card-heading="title"
+            :card-heading="copy.labels.createAccountTitle"
             is-rounded
             has-outline
             is-page-content-wrapper
@@ -11,10 +11,11 @@
             data-test-id="registration-component"
             :class="$style['c-card-padding']">
             <p
-                v-if="shouldShowLoginLink"
+                v-if="showLoginLink"
                 :class="$style['c-loginLink']"
-                data-test-id="create-account-login-link">
-                {{ loginSettings.preLinkText }} <a :href="loginSettings.url">{{ loginSettings.linkText }}</a>
+                data-test-id="create-account-login-link"
+                @click="visitLoginPage">
+                <a :href="copy.navLinks.login.url">{{ copy.navLinks.login.text }}</a>
             </p>
             <form
                 type="post"
@@ -35,7 +36,7 @@
                     v-model="firstName"
                     name="firstName"
                     data-test-id="input-first-name"
-                    label-text="First name"
+                    :label-text="copy.labels.firstName"
                     input-type="text"
                     label-style="inlineNarrow"
                     @blur="formFieldBlur('firstName')">
@@ -45,21 +46,21 @@
                             :class="$style['o-form-error']"
                             data-test-id='error-first-name-empty'>
                             <warning-icon :class="$style['o-form-error-icon']" />
-                            Please include your first name
+                            {{ copy.validationMessages.firstName.requiredError }}
                         </p>
                         <p
                             v-if="shouldShowFirstNameMaxLengthError"
                             :class="$style['o-form-error']"
                             data-test-id='error-first-name-maxlength'>
                             <warning-icon :class="$style['o-form-error-icon']" />
-                            First name exceeds 50 characters
+                            {{ copy.validationMessages.firstName.maxLengthError }}
                         </p>
                         <p
                             v-if="shouldShowFirstNameInvalidCharError"
                             :class="$style['o-form-error']"
                             data-test-id='error-first-name-invalid'>
                             <warning-icon :class="$style['o-form-error-icon']" />
-                            First name should only contain letters, hyphens or apostrophes
+                            {{ copy.validationMessages.firstName.invalidCharError }}
                         </p>
                     </template>
                 </form-field>
@@ -68,7 +69,7 @@
                     v-model="lastName"
                     name="lastName"
                     data-test-id="input-last-name"
-                    label-text="Last name"
+                    :label-text="copy.labels.lastName"
                     input-type="text"
                     label-style="inlineNarrow"
                     @blur="formFieldBlur('lastName')">
@@ -78,21 +79,21 @@
                             :class="$style['o-form-error']"
                             data-test-id='error-last-name-empty'>
                             <warning-icon :class="$style['o-form-error-icon']" />
-                            Please include your last name
+                            {{ copy.validationMessages.lastName.requiredError }}
                         </p>
                         <p
                             v-if="shouldShowLastNameMaxLengthError"
                             :class="$style['o-form-error']"
                             data-test-id='error-last-name-maxlength'>
                             <warning-icon :class="$style['o-form-error-icon']" />
-                            Last name exceeds 50 characters
+                            {{ copy.validationMessages.lastName.maxLengthError }}
                         </p>
                         <p
                             v-if="shouldShowLastNameInvalidCharError"
                             :class="$style['o-form-error']"
                             data-test-id='error-last-name-invalid'>
                             <warning-icon :class="$style['o-form-error-icon']" />
-                            Last name should only contain letters, hyphens or apostrophes
+                            {{ copy.validationMessages.lastName.invalidCharError }}
                         </p>
                     </template>
                 </form-field>
@@ -101,7 +102,7 @@
                     v-model="email"
                     name="email"
                     data-test-id="input-email"
-                    label-text="Email"
+                    :label-text="copy.labels.email"
                     input-type="email"
                     label-style="inlineNarrow"
                     @blur="formFieldBlur('email')">
@@ -111,28 +112,28 @@
                             :class="$style['o-form-error']"
                             data-test-id='error-email-empty'>
                             <warning-icon :class="$style['o-form-error-icon']" />
-                            Please enter your email address
+                            {{ copy.validationMessages.email.requiredError }}
                         </p>
                         <p
                             v-else-if="shouldShowEmailInvalidError"
                             :class="$style['o-form-error']"
                             data-test-id='error-email-invalid'>
                             <warning-icon :class="$style['o-form-error-icon']" />
-                            Please enter a valid email address
+                            {{ copy.validationMessages.email.invalidEmailError }}
                         </p>
                         <p
                             v-if="shouldShowEmailMaxLengthError"
                             :class="$style['o-form-error']"
                             data-test-id='error-email-maxlength'>
                             <warning-icon :class="$style['o-form-error-icon']" />
-                            Email address exceeds 50 characters
+                            {{ copy.validationMessages.email.maxLengthError }}
                         </p>
                         <p
                             v-else-if="shouldShowEmailAlreadyExistsError"
                             :class="$style['o-form-error']"
                             data-test-id='error-email-exists'>
                             <warning-icon :class="$style['o-form-error-icon']" />
-                            Email address is already registered
+                            {{ copy.validationMessages.email.alreadyExistsError }}
                         </p>
                     </template>
                 </form-field>
@@ -141,7 +142,7 @@
                     v-model="password"
                     name="password"
                     data-test-id="input-password"
-                    label-text="Password"
+                    :label-text="copy.labels.password"
                     input-type="password"
                     label-style="inlineNarrow"
                     @blur="formFieldBlur('password')">
@@ -151,14 +152,14 @@
                             :class="$style['o-form-error']"
                             data-test-id='error-password-empty'>
                             <warning-icon :class="$style['o-form-error-icon']" />
-                            Please enter a password
+                            {{ copy.validationMessages.password.requiredError }}
                         </p>
                         <p
                             v-if="shouldShowPasswordMinLengthError"
                             :class="$style['o-form-error']"
                             data-test-id='error-password-minlength'>
                             <warning-icon :class="$style['o-form-error-icon']" />
-                            Password is less than four characters
+                            {{ copy.validationMessages.password.minLengthError }}
                         </p>
                     </template>
                 </form-field>
@@ -168,7 +169,7 @@
                     button-style="primary"
                     is-full-width
                     :disabled="shouldDisableCreateAccountButton">
-                    {{ buttonText }}
+                    {{ copy.labels.createAccountBtn }}
                 </form-button>
             </form>
             <p :class="$style['c-legal-hyperlinks']">
@@ -256,21 +257,13 @@ export default {
             type: String,
             default: 'en-GB'
         },
-        title: {
-            type: String,
-            default: 'Create Account'
-        },
-        buttonText: {
-            type: String,
-            default: 'Create Account'
-        },
         createAccountUrl: {
             type: String,
             required: true
         },
-        loginSettings: {
-            type: Object,
-            default: () => {}
+        showLoginLink: {
+            type: Boolean,
+            default: true
         }
     },
 
@@ -334,9 +327,6 @@ export default {
         shouldShowPasswordMinLengthError () {
             return !this.$v.password.minLength && this.$v.password.$dirty;
         },
-        shouldShowLoginLink () {
-            return this.loginSettings && this.loginSettings.linkText && this.loginSettings.url;
-        },
         tenant () {
             return {
                 'en-GB': 'uk',
@@ -379,6 +369,10 @@ export default {
                 this.$emit(EventNames.CreateAccountStart);
                 this.formStarted = true;
             }
+        },
+
+        visitLoginPage () {
+            this.$emit(EventNames.VisitLoginPage);
         },
 
         formFieldBlur (field) {