+++ packages/f-registration/CHANGELOG.md
@@ -3,6 +3,18 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.35.0
+------------------------------
+*October 16, 2020*
+### Added
+- Moved text into translations file 
+- Support for locale en-GB
+- Emit VisitLoginPage event on login page click interaction
+
+### Changed
+- Prop for showing login link changed to boolean called `showLoginLink`
+
+
 v0.34.0
 ------------------------------
 *October 9, 2020*